requirements.txt
anthropic
apify-client
python-dotenv
requests
jinja2
